+++ pubspec.yaml
@@ -64,6 +64,7 @@ dependencies:
   custom_refresh_indicator: ^2.2.1
   flutter_cache_manager: ^3.3.0
   web_socket_channel: ^2.4.0
+  http_parser: ^4.0.2
 
 dev_dependencies:
   floor_generator: ^1.4.2
